delete alcoComponents folder, move his elements to level up
