+++ client/src/components/AlcoCounter/AlcoCounter.tsx
@@ -3,22 +3,15 @@ import {
   Display,
   AlcoHeader,
   Calendar,
-} from "./alcoComponents";
-import { useAppSelector } from "../../store/hooks/redux";
+} from "./index";
 
 export const AlcoCounter = () => {
-  const { currentTheme } = useAppSelector(
-    (state) => state.userReducer
-  );
-
   return (
-    <>
-      <div className='alco-counter '>
-        <AlcoHeader />
-        <Display />
-        <Calendar />
-        <ControlPanel />
-      </div>
-    </>
+    <section className='alco-counter '>
+      <AlcoHeader />
+      <Display />
+      <Calendar />
+      <ControlPanel />
+    </section>
   );
 };
