+++ client/src/components/AlcoCounter/ControlPanel/ControlPanel.tsx
@@ -1,12 +1,13 @@
 import { Button } from "@mui/material";
-import { ALCO_CONTENT } from "../../../../constants/alcoConstants";
+import { ALCO_CONTENT } from "../../../constants/alcoConstants";
 import { InputLiquidPanel } from "./InputsPanel/InputLiquidPanel";
-import { useAppSelector } from "../../../../store/hooks/redux";
+import { useAppSelector } from "../../../store/hooks/redux";
 
 export const ControlPanel = () => {
   const { currentLang, currentTheme } = useAppSelector(
     (state) => state.userReducer
   );
+  const content = ALCO_CONTENT[currentLang];
   return (
     <div
       className={`control-panel control-panel--${currentTheme}`}
@@ -15,7 +16,7 @@ export const ControlPanel = () => {
       <h2
         className={`control-panel__header control-panel__header--${currentTheme}`}
       >
-        {ALCO_CONTENT[currentLang].controlPanelHeader}
+        {content.controlPanelHeader}
       </h2>
 
       <div className='control-panel__inputs-container'>
