+++ client/src/components/AlcoCounter/ControlPanel/InputsPanel/InputLiquid.tsx
@@ -1,8 +1,8 @@
 import { Button } from "@mui/material";
-import { trimFirstZero } from "../../../../../store/reducer/alcoHandlers";
-import { useAppSelector } from "../../../../../store/hooks/redux";
-import { ALCO_CONTENT } from "../../../../../constants/alcoConstants";
-import { UserThemes } from "../../../../../types/userTypes";
+import { trimFirstZero } from "../../../../store/reducer/alcoHandlers";
+import { useAppSelector } from "../../../../store/hooks/redux";
+import { ALCO_CONTENT } from "../../../../constants/alcoConstants";
+import { UserThemes } from "../../../../types/userTypes";
 
 type InputLiquidProps = {
   val: string;
