+++ client/src/components/AlcoCounter/ControlPanel/InputsPanel/_Input.scss
@@ -1,4 +1,4 @@
-// @import '../../../../../less/variables';
+// @import '../../../../less/variables';
 
 .input-box {
   text-align: center;
