+++ client/src/components/AlcoCounter/ControlPanel/_ControlPanel.scss
