+++ client/src/components/AlcoCounter/Display/_Display.scss
@@ -1,5 +1,3 @@
-// @import'../../../../less/variables';
-
 .display {
   display: flex;
   flex-flow: column wrap;
