--- client/src/components/AlcoCounter/alcoComponents/index.ts
@@ -1,4 +1,0 @@
-export { ControlPanel } from "./ControlPanel/ControlPanel";
-export { Display } from "./Display/Display";
-export { AlcoHeader } from "./AlcoHeader/AlcoHeader";
-export { Calendar } from "./Calendar/Calendar";
