+++ client/src/components/AlcoCounter/index.ts
@@ -1,0 +1,8 @@
+export { AlcoHeader } from "./AlcoHeader/AlcoHeader";
+export { Calendar } from "./Calendar/Calendar";
+
+export { ControlPanel } from "./ControlPanel/ControlPanel";
+
+export { Display } from "./Display/Display";
+
+export { AlcoCounter } from "./AlcoCounter";
